Откорректирован тест GameFactory, добавлена проверка сообщения исключения.
